requirements.txt
numpy==1.19.4
pygame==2.0.0
PyOpenGL==3.1.5
